third_party/WebKit/LayoutTests/fast/html/imports/import-document-write.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="import" href="resources/do-document-write.html">
</head>
<body>
<x-test></x-test>
<script>
test(function(){
    var importedDocument = document.querySelector("link").import;
    assert_throws("InvalidStateError", function() { importedDocument.open() }, "open() doesn't work for imports");
    assert_throws("InvalidStateError", function() { importedDocument.write("Hello") }, "write() doesn't work for imports");
    assert_throws("InvalidStateError", function() { importedDocument.close() }, "close() doesn't work for imports");
}, "The document write and family isn't supported.");
</script>
</body>
</html>
